fix: #1917 Purge unused code from login page
