+++ packages/developer-portal/src/components/pages/login/index.ts
@@ -1,0 +1,2 @@
+import Login from './login'
+export default Login
